refactor: extract AccountBillingApi from accounts.rs (#444)
