+++ api/src/openapi.rs
@@ -1,4 +1,5 @@
 pub mod accounts;
+pub mod accounts_billing;
 pub mod accounts_contacts;
 pub mod accounts_keys;
 pub mod admin;
@@ -34,6 +35,7 @@ pub mod webhooks;
 pub mod webhooks_disputes;
 
 pub use accounts::AccountsApi;
+pub use accounts_billing::AccountBillingApi;
 pub use accounts_contacts::AccountContactsApi;
 pub use accounts_keys::AccountKeysApi;
 pub use admin::AdminApi;
@@ -87,6 +89,7 @@ pub fn create_combined_api() -> impl OpenApi {
             SlaApi,
             AllowlistApi,
             ContractTelemetryApi,
+            AccountBillingApi,
         ),
         (
             OfferingsApi,
