+++ api/src/openapi/accounts.rs
@@ -620,102 +620,6 @@ impl AccountsApi {
         }
     }
 
-    /// Get billing settings
-    ///
-    /// Returns saved billing information for the authenticated user
-    #[oai(path = "/accounts/billing", method = "get", tag = "ApiTags::Accounts")]
-    async fn get_billing_settings(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-    ) -> Json<ApiResponse<crate::database::accounts::BillingSettings>> {
-        // Get account by authenticated user's public key
-        let account_id = match db.get_account_id_by_public_key(&auth.pubkey).await {
-            Ok(Some(id)) => id,
-            Ok(None) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some("Authenticated key not found or not active".to_string()),
-                })
-            }
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e.to_string()),
-                })
-            }
-        };
-
-        // Get billing settings
-        match db.get_billing_settings(&account_id).await {
-            Ok(settings) => Json(ApiResponse {
-                success: true,
-                data: Some(settings),
-                error: None,
-            }),
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e.to_string()),
-            }),
-        }
-    }
-
-    /// Update billing settings
-    ///
-    /// Updates saved billing information for the authenticated user
-    #[oai(path = "/accounts/billing", method = "put", tag = "ApiTags::Accounts")]
-    async fn update_billing_settings(
-        &self,
-        db: Data<&Arc<Database>>,
-        auth: ApiAuthenticatedUser,
-        req: Json<crate::database::accounts::BillingSettings>,
-    ) -> Json<ApiResponse<crate::database::accounts::BillingSettings>> {
-        // Get account by authenticated user's public key
-        let account_id = match db.get_account_id_by_public_key(&auth.pubkey).await {
-            Ok(Some(id)) => id,
-            Ok(None) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some("Authenticated key not found or not active".to_string()),
-                })
-            }
-            Err(e) => {
-                return Json(ApiResponse {
-                    success: false,
-                    data: None,
-                    error: Some(e.to_string()),
-                })
-            }
-        };
-
-        // Update billing settings
-        match db.update_billing_settings(&account_id, &req.0).await {
-            Ok(_) => {
-                // Fetch updated settings
-                match db.get_billing_settings(&account_id).await {
-                    Ok(settings) => Json(ApiResponse {
-                        success: true,
-                        data: Some(settings),
-                        error: None,
-                    }),
-                    Err(e) => Json(ApiResponse {
-                        success: false,
-                        data: None,
-                        error: Some(e.to_string()),
-                    }),
-                }
-            }
-            Err(e) => Json(ApiResponse {
-                success: false,
-                data: None,
-                error: Some(e.to_string()),
-            }),
-        }
-    }
     /// Delete my account
     ///
     /// Permanently deletes the authenticated account and all associated data.
@@ -805,7 +709,7 @@ impl AccountsApi {
 
 #[cfg(test)]
 mod tests {
-    use crate::database::accounts::{AccountWithKeys, BillingSettings, PublicKeyInfo};
+    use crate::database::accounts::{AccountWithKeys, PublicKeyInfo};
     use crate::openapi::common::{
         ApiResponse, RegisterAccountRequest, UpdateAccountEmailRequest, UpdateAccountProfileRequest,
     };
@@ -946,34 +850,6 @@ mod tests {
         let json = serde_json::to_value(&resp).unwrap();
         assert_eq!(json["success"], false);
         assert!(json["error"].as_str().unwrap().contains("Unauthorized"));
-    }
-
-    // ---- BillingSettings ----
-
-    #[test]
-    fn test_billing_settings_serialization_camel_case() {
-        let settings = BillingSettings {
-            billing_address: Some("123 Main St".to_string()),
-            billing_vat_id: Some("VAT123".to_string()),
-            billing_country_code: Some("US".to_string()),
-        };
-        let json = serde_json::to_value(&settings).unwrap();
-        assert_eq!(json["billingAddress"], "123 Main St");
-        assert_eq!(json["billingVatId"], "VAT123");
-        assert_eq!(json["billingCountryCode"], "US");
-    }
-
-    #[test]
-    fn test_billing_settings_all_none() {
-        let settings = BillingSettings {
-            billing_address: None,
-            billing_vat_id: None,
-            billing_country_code: None,
-        };
-        let json = serde_json::to_value(&settings).unwrap();
-        assert!(json.get("billingAddress").is_none());
-        assert!(json.get("billingVatId").is_none());
-        assert!(json.get("billingCountryCode").is_none());
     }
 
     // ---- Validation functions (called inline in handler, tested here) ----
